requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
cffi==2.0.0
clerk-backend-api==4.0.0
click==8.3.0
colorama==0.4.6
cryptography==45.0.7
dnspython==2.8.0
email-validator==2.3.0
et-xmlfile==2.0.0
fastapi==0.117.1
fastapi-clerk-auth==0.0.9
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
pillow==12.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
qrcode==8.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.48.0
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.0
websockets==15.0.1
